WIP moving MXF utility methods out of IMPValidator/Analyzer into new MXFUtils class
